apps/api-server/pyproject.toml
[project]
name = "api-server"
version = "0.1.0"
description = "FastAPI API server for MiroSearch, independent from Gradio demo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "hydra-core>=1.3.0",
    "miroflow-agent",
    "arq>=0.26.0",
    "redis>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["./"]

[tool.uv.sources]
miroflow-agent = { path = "../miroflow-agent", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
minversion = "8.3.5"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
